Added status and poncho command, grouped commands with a random pix or text
