+++ nerdlandbot/eventhandlers/onready.py
@@ -1,4 +1,5 @@
 from discord.ext import commands
+import discord 
 
 from nerdlandbot.helpers.log import info
 
@@ -13,7 +14,7 @@ class OnReady(commands.Cog, name="on_ready"):
         This gets executed when the bot connects to discord.
         """
         info(f'{self.bot.user.name} has connected to Discord!')
-
+        await self.bot.change_presence(activity=discord.Game(name=f"with {self.bot.command_prefix}poncho"))
 
 def setup(bot: commands.Bot):
     bot.add_cog(OnReady(bot))
