+++ nerdlandbot/translations/Translations.csv
@@ -116,6 +116,7 @@
 "poll_brief","Put a poll out for people to vote","Maak een peiling waarin mensen kunnen stemmen"
 "poll_help","Start a poll, with minimal arguments this will be a yes/no poll, if you provide options seperated by semicolons you can add up to 9 options.","Start een peiling, met minimale opties wordt dit een Ja/Nee peiling die <timeout> minuten duurt. Indien je zelf opties meegeeft door puntkomma gescheiden kunnen dat er maximaal 9 zijn."
 "empty_wombat_list","No wombat pictures were found...","Er werden geen foto's van wombats gevonden..."
+"wombat_help","Posts a random wombat picture.","Post een willekeurige foto van een wombat."
 "add_purger_usage","<Channel ID> <Max message age (in days)>","<Kanaal ID> <Max leeftijd bericht (in dagen)>"
 "add_purger_help","**admin-only** \n \u2003 Automatically remove messages older than the configured amount of days in the channel.","**admin-functie** \n \u2003 Verwijder automatisch berichten die ouder zijn dan het opgegeven aantal dagen in het kanaal."
 "remove_purger_usage","<Channel ID>","<Kanaal ID>"
@@ -139,3 +140,6 @@
 "privacy_brief","Prints the privacy policy.","Toont de privacy policy die door de bot gehanteerd wordt."
 "privacy_policy_title","Privacy Policy","Privacy Policy"
 "privacy_policy_content","This bot was developed with privacy as one of our core ideals.\nBecause of this we have formulated a few statements about the inner workings.\n\nThe bot will only listen to user commands.\nWe will not parse any user messages, nor track any reactions, unless those made specifically to interact with the bot.\nWe will not track any user data, except for your user ID, as this is required to notify you.\nWe will not store any data about your messages, reactions, or other actions you take on discord.\n\nThe logs we keep are strictly for debugging purposes, and will not contain any personal info.","Bij de ontwikkeling van deze bot hebben we privacy bekeken als een zeer belangrijk punt, om die reden hebben we onze mening hierover uitgedrukt in enkele punten.\n\nDe bot luistert enkel naar commando's die beginnen met de bepaalde prefix.\nDe bot zal nooit berichten die niet met de gegeven prefix beginnen parsen, ook houden we geen reacties bij, behalve de reacties op berichten van de bot.\nDe bot houdt geen gebruikersdata bij, behalve het gebruikers-ID, aangezien dit nodig is om gebruikers te taggen.\nDe bot houdt geen data bij over je berichten, reacties, of andere acties op discord.\n\nDe logs die we bijhouden worden enkel gebruikt om te debuggen. Deze bevatten geen persoonlijke data."
+"empty_poncho_list","No Poncho pictures were found...","Er werden geen foto's van Poncho gevonden..."
+"poncho_help","Posts a random Poncho picture.","Post een willekeurige foto van Poncho."
+"poncho_hond","DOG DOG DOG","HOND HOND HOND"
